Consolidate search-input and search-header
